feat: clean up custom path components
