+++ apps/extension/src/Setup/Common/Zip32Form.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { chains } from "@namada/chains";
-import { Alert, Input, LinkButton, Stack } from "@namada/components";
+import { Input } from "@namada/components";
 import { Zip32Path } from "@namada/types";
 
 type Props = {
@@ -26,60 +26,24 @@ const Zip32Form: React.FC<Props> = ({ path, setPath }) => {
 
   const parentDerivationPath = `m/32'/${coinType}'/`;
 
-  const handleReset = (): void => {
-    setPath({ account: 0 });
-  };
-
   return (
-    <div className="flex flex-col w-full">
-      <div className="my-3">
-        <Alert
-          type="info"
-          title="Please note"
-          className="[&_strong]:normal-case"
-        >
-          <Stack gap={6}>
-            <ul className="text-white list-disc mx-6 [&_button]:text-white">
-              <li>
-                You can create multiple addresses from one recovery phrase
-              </li>
-              <li>A lost path cannot be recovered</li>
-              <li>
-                If you&apos;re unfamiliar with this feature, skip or undo this
-                step -{" "}
-                <LinkButton
-                  className="underline bold"
-                  role="button"
-                  onClick={(e) => {
-                    e.preventDefault();
-                    handleReset();
-                  }}
-                >
-                  Reset
-                </LinkButton>
-              </li>
-            </ul>
-            <label className="text-base font-medium text-neutral-300">
-              ZIP32 Path
-              <div className="flex w-full justify-start items-center pt-2">
-                <span className="h-px px-1 text-xs text-neutral-300">
-                  {parentDerivationPath}
-                </span>
-                <Input
-                  type="number"
-                  className="w-60"
-                  min="0"
-                  value={path.account}
-                  onChange={handleNumericChange}
-                  onFocus={handleFocus}
-                />
-                <i>&apos;</i>
-              </div>
-            </label>
-          </Stack>
-        </Alert>
+    <label className="text-base font-medium text-neutral-300">
+      ZIP32 Path
+      <div className="flex w-full justify-start items-center pt-2">
+        <span className="h-px px-1 text-xs text-neutral-300">
+          {parentDerivationPath}
+        </span>
+        <Input
+          type="number"
+          className="w-60"
+          min="0"
+          value={path.account}
+          onChange={handleNumericChange}
+          onFocus={handleFocus}
+        />
+        <i>&apos;</i>
       </div>
-    </div>
+    </label>
   );
 };
